requirements.txt
bcrypt
cloudinary
fastapi[all]
surrealdb
python-jose
prisma;python_version>='3.11'

